#316: enhancement: Refacto
